Implement and register `sd plan ready` (status !== done)
Add a 'sd plan ready' subcommand that lists plans whose status is not
'done' (draft|approved|active), mirroring 'sd plan list' filters and
output. Exposes --seed/--outcome/--template/--json passthrough filters
(no --status, since not-done is the verb's purpose).

Extracted the shared plan-list helpers (ListFilters, filter/render/
runPlanList) from src/commands/plan.ts into src/commands/plan-list.ts
to keep plan.ts under the file-size ratchet budget.

Closes seeds-25de (pl-611d step 2).
