+++ .mulch/expertise/commands.jsonl
@@ -77,3 +77,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-29T01:19:45.657Z","evidence":{"seeds":"seeds-3dd1","commit":"0fae93b593c602a02a114292705fa781b7b31d0d"},"dir_anchors":["src/commands"],"content":"When plan.ts crosses the check:size budget (scripts/file-size-budgets.json), extract a cohesive block into a new module rather than raising the budget (ratchet only goes down). The plan-show render/tree helpers + runShow + resolvePlanIdArg moved to src/commands/plan-show.ts; plan.ts re-imports runShow/resolvePlanIdArg and src/commands/show.ts imports runShow from there. New files needing the relaxed cognitive-complexity rule must be added to the includes[] list in biome.json's overrides (e.g. plan-show.ts alongside plan.ts).","id":"mx-1d2f20"}
 {"type":"pattern","classification":"tactical","recorded_at":"2026-06-16T10:28:34.696Z","evidence":{"seeds":"seeds-3024","commit":"032374e6202cf3fd2ba22d2a01753998627b12c6"},"dir_anchors":["src"],"name":"epipe-safe-stdout","description":"seeds CLI handles broken-pipe (sd ... --json | head) two ways: (1) process.stdout/stderr .on('error') EPIPE handlers in src/index.ts exit(0) for console.log/process.stdout.write output; (2) the Bun.write(Bun.stdout) path bypasses those stream objects, so src/output.ts exports writeStdout() which try/catches EPIPE and process.exit(0). outputJson + all Bun.write(Bun.stdout) sites (index.ts unknown-command branch, error-handler.ts, config.ts) route through writeStdout. Exit 0 = downstream hangup is not an sd failure. Without this a large write to a closed pipe throws EPIPE / busy-spins on Linux.","files":[".seeds/issues.jsonl","src/cli-smoke.test.ts","src/commands/config.ts","src/error-handler.ts","src/index.ts","src/output.ts"],"id":"mx-b1c05f"}
 {"type":"pattern","classification":"tactical","recorded_at":"2026-06-20T19:18:02.687Z","evidence":{"commit":"c8cc90f74b1449b3009865ecd8699af464a2aa27"},"name":"plan-list-shared-render","description":"src/commands/plan.ts splits runList into validateListFilters (enum checks) + applyListFilters (positive --seed/--status/--outcome/--template filter + newest-first sort) + renderPlanList(filtered, command, jsonMode) (shared human+JSON output). 'plan ready' reuses these, pre-narrowing status!=='done' before applyListFilters, passing 'plan ready' as the JSON command field (seeds-03f2 / pl-611d step 1).","files":[".canopy/agent.json",".seeds/issues.jsonl",".seeds/plans.jsonl","src/commands/plan.ts"],"id":"mx-221430"}
+{"type":"pattern","classification":"tactical","recorded_at":"2026-06-20T21:31:35.727Z","evidence":{"seeds":"seeds-25de","commit":"c60ef9c8e20bb6355e74448640ff1970124f6820"},"name":"plan-ready-command","description":"sd plan ready (src/commands/plan-list.ts runPlanList, seeds-25de/pl-611d): mirrors sd plan list but pre-narrows status!=='done' (notDone=true) before applyListFilters, exposing only --seed/--outcome/--template/--json (no --status). list/ready share runPlanList(command, filters, jsonMode, notDone); command keeps the JSON 'command' field distinct. Plan-list helpers were moved out of plan.ts into plan-list.ts to stay under the file-size ratchet budget.","files":[".seeds/issues.jsonl",".seeds/plans.jsonl","src/commands/plan.ts"],"id":"mx-f9d4e1"}
